Consolidate profile session code
