+++ nova_lxd/nova/virt/lxd/session/session.py
@@ -28,7 +28,6 @@ from oslo_utils import excutils
 
 from nova_lxd.nova.virt.lxd import constants
 from nova_lxd.nova.virt.lxd.session import migrate
-from nova_lxd.nova.virt.lxd.session import profile
 from nova_lxd.nova.virt.lxd.session import snapshot
 
 _ = i18n._
@@ -41,7 +40,6 @@ LOG = logging.getLogger(__name__)
 
 
 class LXDAPISession(migrate.MigrateMixin,
-                    profile.ProfileMixin,
                     snapshot.SnapshotMixin):
     """The session to invoke the LXD API session."""
 
@@ -660,3 +658,96 @@ class LXDAPISession(migrate.MigrateMixin,
                               '%(instance)s: %(reason)s'),
                           {'instance': instance.image_ref, 'reason': e},
                           instance=instance)
+
+    #
+    # Profile methods
+    #
+    def profile_defined(self, instance):
+        """Validate if the profile is available on the LXD
+           host
+
+           :param instance: nova instance object
+        """
+        LOG.debug('profile_defined called for instance',
+                  instance=instance)
+        try:
+            client = self.get_session(instance.host)
+            client.profile_defined(instance.name)
+        except lxd_exceptions.APIError as ex:
+            if ex.status_code == 404:
+                return False
+            else:
+                msg = _('Failed to communicate with LXD API %(instance)s:'
+                        ' %(reason)s') % {'instance': instance.name,
+                                          'reason': ex}
+                raise exception.NovaException(msg)
+        except Exception as ex:
+            with excutils.save_and_reraise_exception():
+                LOG.error(
+                    _LE('Failed to determine profile %(instance)s:'
+                        ' %(reason)s'),
+                    {'instance': instance.name, 'reason': ex})
+
+    def profile_create(self, config, instance):
+        """Create an LXD container profile
+
+        :param config: profile dictionary
+        :param instnace: nova instance object
+        """
+        LOG.debug('profile_create called for instance',
+                  instance=instance)
+        try:
+            client = self.get_session(instance.host)
+            client.profile_create(config)
+        except lxd_exceptions.APIError as ex:
+            msg = _('Failed to communicate with LXD API %(instance)s:'
+                    ' %(reason)s') % {'instance': instance.name,
+                                      'reason': ex}
+            raise exception.NovaException(msg)
+        except Exception as ex:
+            with excutils.save_and_reraise_exception():
+                LOG.error(
+                    _LE('Failed to create profile %(instance)s: %(reason)s'),
+                    {'instance': instance.name, 'reason': ex})
+
+    def profile_update(self, config, instance):
+        """Update an LXD container profile
+
+          :param config: LXD profile dictironary
+          :param instance: nova instance object
+        """
+        LOG.debug('profile_udpate called for instance', instance=instance)
+        try:
+            client = self.get_session(instance.host)
+            client.profile_update(instance.name, config)
+        except lxd_exceptions.APIError as ex:
+            msg = _('Failed to communicate with LXD API %(instance)s:'
+                    ' %(reason)s') % {'instance': instance.name,
+                                      'reason': ex}
+            raise exception.NovaException(msg)
+        except Exception as ex:
+            with excutils.save_and_reraise_exception():
+                LOG.error(
+                    _LE('Failed to update profile %(instance)s: '
+                        '%(reason)s'),
+                    {'instance': instance.name, 'reason': ex})
+
+    def profile_delete(self, instance):
+        """Delete a LXD container profile.
+
+           :param instance: nova instance object
+        """
+        LOG.debug('profile_delete called for instance', instance=instance)
+        try:
+            client = self.get_session(instance.host)
+            client.profile_delete(instance.name)
+        except lxd_exceptions.APIError as ex:
+            msg = _('Failed to communicate with LXD API %(instance)s:'
+                    ' %(reason)s') % {'instance': instance.name,
+                                      'reason': ex}
+            raise exception.NovaException(msg)
+        except Exception as ex:
+            with excutils.save_and_reraise_exception():
+                LOG.error(
+                    _LE('Failed to delete profile %(instance)s: %(reason)s'),
+                    {'instance': instance.name, 'reason': ex})
